move padding model wrapper under example
